Create support-test-fakes component and move FakeEngine implementation to it
