+++ .buildconfig.yml
@@ -327,6 +327,10 @@ projects:
     path: components/support/test
     description: 'A collection of helpers for testing components (local unit tests).'
     publish: true
+  support-test-fakes:
+    path: components/support/test-fakes
+    description: 'A collection of fake implementations for testing purposes.'
+    publish: true
   support-test-libstate:
     path: components/support/test-libstate
     description: 'A collection of helpers for testing functionality that relies on the lib-state component (local unit tests).'
