+++ components/browser/session-storage/build.gradle
@@ -47,6 +47,7 @@ dependencies {
     implementation Dependencies.androidx_lifecycle_process
 
     testImplementation project(':support-test')
+    testImplementation project(':support-test-fakes')
     testImplementation project(':support-test-libstate')
     testImplementation project(':feature-tabs')
     testImplementation Dependencies.androidx_test_core
