+++ components/feature/recentlyclosed/build.gradle
@@ -71,6 +71,7 @@ dependencies {
     testImplementation Dependencies.kotlin_coroutines
     testImplementation Dependencies.testing_coroutines
 
+    androidTestImplementation project(":support-test-fakes")
     androidTestImplementation project(':support-android-test')
     androidTestImplementation Dependencies.androidx_room_testing
     androidTestImplementation Dependencies.androidx_arch_core_testing
