+++ components/feature/session/src/test/java/mozilla/components/feature/session/behavior/EngineViewBrowserToolbarBehaviorTest.kt
@@ -5,15 +5,13 @@
 package mozilla.components.feature.session.behavior
 
 import android.content.Context
-import android.graphics.Bitmap
 import android.view.View
 import android.widget.EditText
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.EngineView
-import mozilla.components.concept.engine.selection.SelectionActionDelegate
+import mozilla.components.support.test.fakes.engine.FakeEngineView
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.testContext
 import org.junit.Assert.assertEquals
@@ -153,22 +151,6 @@ class EngineViewBrowserToolbarBehaviorTest {
 
         verify(engineView).setVerticalClipping(41)
     }
-}
-
-class FakeEngineView(context: Context) : TextView(context), EngineView {
-    override fun render(session: EngineSession) {}
-
-    override fun captureThumbnail(onFinish: (Bitmap?) -> Unit) {}
-
-    override fun clearSelection() {}
-
-    override fun setVerticalClipping(clippingHeight: Int) {}
-
-    override fun setDynamicToolbarMaxHeight(height: Int) {}
-
-    override fun release() {}
-
-    override var selectionActionDelegate: SelectionActionDelegate? = null
 }
 
 class BrowserToolbar(context: Context) : TextView(context)
